requirements.txt
Flask[async]>=2.3
gunicorn
python-telegram-bot>=22.0
python-dotenv
google-genai

